👷 Simplify prod env vars
Instead of putting the .env file in some other place with a different name, just use .env in docker locally and then copy .prod.env over as .env on deploy.

This was necessary/easier to do because the `env_file` option still gets evaluated in all places, even when docker-compose is invoked in a way that's meant to override it. So trying to point to `.env` in docker-compose.yml and then a different file in docker-compose.prod.yml caused it to fail because it couldn't find .env. Just making it so the .env is where we need it in all environments resolves this and is way simpler.